\name{\%thru\%}
\alias{\%thru\%}

\title{
Funcion auxiliar logica para valores dentro de un rango
}

\description{
Funcion similiar a la de between del paquete data.table, que permite verificar de manera logica si un valor pertenece a un rango de valores
}
\usage{
thru(x,inicial,final)
x \%thru\% y
}

\arguments{
  \item{x}{
Vector numerico
}
  \item{inicial}{
Limite inferior del rango
}
  \item{final}{
Limite superior del rango
}
 \item{y}{
Vector de longitud 2, que indique el limite inferior y el limite superior del rango
}
}

\author{
Emilio Arenas
}

\keyword{thru}
