\name{addvallab}
\alias{addvallab}

\title{
Añadir etiquetas de valor
}
\description{
Funcion que permite añadir definiciones de valores a una variable con valores ya definidos
}
\usage{
addvallab(df,var,val)
}

\arguments{
  \item{df}{
Objeto del tipo data.frame donde se encuentra la variable cuyas definiciones de valor se desean ampliar
}
  \item{var}{
La variable en cuestion
}
  \item{val}{
Vector con nombres, por ejemplo c("Si"=1,"No"=2)
}
}

\author{
Emilio Arenas
}

\keyword{addvallab}