\name{compnumeric}
\alias{compnumeric}

\title{
Transformar un factor en variable numerica categorica
}
\description{
Funcion que transforma un factor en variable numerica categorica utilizando sus niveles como attr(x,"val.lab")
}
\usage{
compnumeric(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Una variable Factor
}
}

\value{
Devuelve una variable numerica categorica con attr(x,"val.lab")
}

\author{
Emilio Arenas
}

\keyword{compnumeric}