\name{contar}
\alias{contar}

\title{
Conteo
}

\description{
Funcion que devuelve para cada observacion el numero de ocurrencias que coincide con un valor logico
}
\usage{
contar(bbff, v, thru=F)
}

\arguments{
  \item{bbdd}{
Data.frame
}
  \item{v}{
Valor o vector de valores que se desea contar
}
  \item{thru}{
Por defecto FALSE, si v es de longitud 2 y thru=TRUE, entonces devolvera el numero de ocurrencias dentro del rango v
}
}

\author{
Emilio Arenas
}

\keyword{contar}
