\name{correl}
\alias{correl}

\title{
Matriz de correlaciones
}
\description{
Esta funcion calcula la matriz de correlaciones de Pearson
}
\usage{
correl(x,w)
}

\arguments{
  \item{x}{
Matrix o data.frame de 2 o mas columnnas que contenga variables numericas
}
  \item{w}{
Variable de ponderacion de los casos
}
}

\author{
Emilio Arenas
}

\keyword{correl}
