\name{count}
\alias{count}

\title{
Conteo
}

\description{
Funcion que devuelve para cada observacion el numero de ocurrencias que coincide con un valor logico
}
\usage{
count(..., v)
}

\arguments{
  \item{...}{
Variables
}
  \item{v}{
Valor o vector de valores que se desea contar
}
}

\author{
Emilio Arenas
}

\keyword{count}