\name{covar}
\alias{covar}

\title{
Matriz de covarianzas
}
\description{
Esta funcion calcula la matriz de covarianzas
}
\usage{
covariance(x,w)
}

\arguments{
  \item{x}{
Matrix o data.frame de 2 o mas columnnas que contenga variables numericas
}
  \item{w}{
Variable de ponderacion de los casos
}
}

\author{
Emilio Arenas
}

\keyword{covar}
