\name{export.sav}
\alias{export.sav}

\title{
Exportar de data.frames a SAV
}
\description{
Funcion que permite pasar data.frames a SAV
}
\usage{
export.sav(df, path)
}

\arguments{
  \item{df}{
Data.frame
}
  \item{path}{
Nombre y ubicacion del archivo de datos
}
}
\author{
Emilio Arenas
}
\keyword{export.sav}
