\name{filtro}
\alias{filtro}

\title{
Filtrado de bases de datos
}
\description{
Funcion que permite seleccionar casos y variables de un data.frame sin perder las definiciones
}
\usage{
filtro(df, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{df}{
Data.frame
}
  \item{\dots}{
argumentos basados en la funcion subset( )
}
}

\value{
Devuelve un data.frame
}

\author{
Emilio Arenas
}

\keyword{filtro}