\name{list.val}
\alias{list.val}

\title{
Listar Value Labels
}
\description{
Funcion que permite visualizar los value labels de cualquier variable
}
\usage{
list.val(x)
}

\arguments{
  \item{x}{
Variable
}
}

\value{
Data.frame que contiene los value labels
}

\author{
Emilio Arenas
}

\keyword{list.val}