\name{list.var}
\alias{list.var}

\title{
Listar Variables
}
\description{
Funcion que permite buscar variables dentro de un data.frame
}
\usage{
list.var(str, df)
}

\arguments{
  \item{str}{
Valor string que deseamos buscar en df
}
  \item{df}{
Data.frame donde realizaremos las busqeda de str
}
}

\value{
Data.frame con las variables y sus deficiones que coincidan en la busqueda
}

\author{
Emilio Arenas
}

\keyword{list.var}