\name{print.comul}
\alias{print.comul}

\title{
Imprime por pantalla el resultado del analisis de correspondencias simple de K-tablas
}
\description{
Imprime por pantalla el resultado del analisis de correspondencias simple de K-tablas
}
\usage{
print(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Objeto de la clase comul
}
  \item{\dots}{
argumentos basados en la funcion print.CA( )
}
}

\value{
No devuelve ningun objeto
}

\author{
Emilio Arenas
}

\keyword{print.comul}