\name{print.tabla}
\alias{print.tabla}

\title{
Outputs de matrix de clases tabla
}
\description{
Funcion que permite obtener un output similar al SPSS, para la funciones del paquete estadisticos
}
\usage{
print.tabla(x)
}

\arguments{
  \item{x}{
Objeto de la clase c("tabla","matrix")
}
}


\author{
Emilio Arenas
}

\keyword{print.tabla}