\name{spssdef}
\alias{spssdef}

\title{
Definir data.frame con las deficiones de un archivo SAV
}
\description{
Funcion que permite definir un data.frame obtenido a partir de la funcion read.spss(file,use.value.labels=F,to.data.frame=T) del paquete foreign
}
\usage{
spssdef(df)
}
\arguments{
  \item{df}{
Data.frame
}
}

\value{
Devuelve un data.frame
}

\author{
Emilio Arenas
}

\keyword{spssdef}