\name{standarized}
\alias{standarized}

\title{
Standarizar Variables
}
\description{
Funcion que obtener variables standarizadas
}
\usage{
standarized(df, w)
}

\arguments{
  \item{df}{
Data.frame con las variables a standarizar
}
  \item{w}{
Variable de poderacion
}
}

\value{
Data.frame con variables de Media=0 y Desviacion estandar=1
}

\author{
Emilio Arenas
}

\keyword{standarized}