\name{sumar}
\alias{sumar}

\title{
Sumatorio
}
\description{
Suma horizontal (para cada caso) de una serie de variables
}
\usage{
sumar(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{...}{
lista de variables
}
}

\value{
Devuelve un vector de la misma longitud que la lista de variables
}
\references{
Funcion rowSums
}
\author{
Emilio Arenas
}

\keyword{sumar}