\name{summary.comul}
\alias{summary.comul}

\title{
Resumen del analisis de correspondencias simple de K-tablas
}
\description{
Funcion que permite visualizar de un modo más "amigable" los resultado del analisis de correspondencias simple de K-tablas
}
\usage{
summary(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Objeto de la clase comul
}
  \item{\dots}{
argumentos basados en la funcion summary.CA( )
}
}

\value{
No devuelve ningun objeto
}

\author{
Emilio Arenas
}

\keyword{summary.comul}