\name{thr}
\alias{thr}
\alias{\%thr\%}

\title{
Funcion auxiliar logica para valores dentro de un rango
}

\description{
Funcion similiar a la de between del paquete data.table, que permite verificar de manera logica si un valor pertenece a un rango de valores
}
\usage{
thr(x,inicial,final)
x \%thr\% y
}

\arguments{
  \item{x}{
Vector numerico
}
  \item{inicial}{
Limite inferior del rango
}
  \item{final}{
Limite superior del rango
}
}

\author{
Emilio Arenas
}

\keyword{thr}
