\name{vallab}
\alias{vallab}

\title{
Etiquetas de valor
}
\description{
Funcion que permite definir los codigos de valor de las variables numericas
}
\usage{
vallab(df,var,val)
}

\arguments{
  \item{df}{
Objeto del tipo data.frame donde se encuentra la variable cuyos valores se quieren definir
}
  \item{var}{
La variable en cuestion
}
  \item{val}{
Vector con nombres, por ejemplo c("Si"=1,"No"=2)
}
}

\author{
Emilio Arenas
}

\keyword{vallab}