\name{varlab}
\alias{varlab}

\title{
Etiqueta de variable
}
\description{
Funcion que permite definir variables
}
\usage{
varlab(df,var,val)
}

\arguments{
  \item{df}{
Objeto del tipo data.frame donde se encuentra la variable que se desea definir
}
  \item{var}{
La variable en cuestion
}
  \item{val}{
La definicion de la variable introducida como texto entrecomillado.
}
}

\author{
Emilio Arenas
}

\keyword{varlab}