\name{contar}
\alias{contar}

\title{
Conteo
}

\description{
Funcion que devuelve para cada observacion el numero de ocurrencias que coincide con un valor logico
}
\usage{
count(bbff, v,thr=F)
}

\arguments{
  \item{bbdd}{
Data.frame
}
  \item{v}{
Valor o vector de valores que se desea contar
}
\item{thr}{
Por defecto FALSE, si v es de longitud 2 y thr=TRUE, entonces devolvera el numero de ocurrencias dentro del rango v
}
}

\author{
Emilio Arenas
}

\keyword{count}
